Nonlinear Chatbot
After months of deliberating i found the time and energy to dive into creating a nonlinear mima draft. Or at least a draft where the bot is more responsive and dynamic. I took out the background music for now as I'm writing it for people using Mima for the larp and who may be playing the LP at the same time.